+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 0.13.0
-- `@use-gpu/scene`: Fix instancing inheritance
+- `@use-gpu/scene`: Support pickable `<Instances>` with per instance `lookup`
 - `@use-gpu/wgsl-loader`: Pass `minify` option to rollup plug-in.
 - `@use-gpu/workbench`: Fix JSX type of <Dispatch> and <DrawCall> for React 18
 - `@use-gpu/workbench`: Allow `format`-less <RenderTarget>